requirements.txt
blinker==1.9.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cryptography==43.0.3
cycler==0.12.1
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dash_daq==0.5.0
Flask==3.0.3
fonttools==4.54.1
gunicorn==23.0.0
idna==3.10
importlib_metadata==8.5.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.2
nest-asyncio==1.6.0
numpy==2.1.1
opencv-python==4.10.0.84
packaging==24.2
paho-mqtt==2.1.0
pillow==10.4.0
plotly==5.24.1
pluggy==1.5.0
pycparser==2.22
PyJWT==2.10.0
pyparsing==3.1.4
pytest==8.3.3
python-dateutil==2.9.0.post0
pytz==2024.2
pyzbar==0.1.9
requests==2.32.3
retrying==1.3.4
setuptools==75.5.0
six==1.16.0
tenacity==9.0.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.6
zipp==3.21.0
